requirements.txt
motor==3.3.2  # Use the latest version available
fastapi==0.103.2
uvicorn==0.23.2
httpx==0.25.0
python-dotenv==1.0.0
pydantic==2.4.2
asyncio==3.4.3



